Better UI

+++ pages/customer/index.tsx
@@ -1,25 +1,25 @@
-import Button from "@mui/material/Button";
-import Grid from "@mui/material/Grid";
-import Container from "@mui/material/Container";
-import TextField from "@mui/material/TextField";
-import PostModal from "@/components/Customer/CustomerModal";
-import { useEffect, Fragment, useState, ChangeEvent } from "react";
+import {
+  Box,
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TablePagination,
+  TableRow,
+  Paper,
+  Checkbox,
+  LinearProgress,
+} from "@mui/material";
 
-import OptionsButton from "@/components/OptionsButton";
-import List from "@mui/material/List";
-import Divider from "@mui/material/Divider";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-import InboxIcon from "@mui/icons-material/MoveToInbox";
-import CreateIcon from "@mui/icons-material/Create";
-import DeleteIcon from "@mui/icons-material/Delete";
-import CircularProgress from "@mui/material/CircularProgress";
-import { Typography } from "@mui/material";
-
-import CustomerCard from "@/components/Customer/CustomerCard";
-import Box from "@mui/material/Box";
+import React, { useEffect, useState, ChangeEvent } from "react";
+import { useDispatch } from "react-redux";
+import { useRouter } from "next/router";
+import {
+  processRequest,
+  processRequestNonReponse,
+  sendMessageAction,
+  getHeaders,
+} from "@/pages/index";
 
 import {
   breadcrumbAction,
@@ -27,42 +27,92 @@ import {
   CUSTOMER,
 } from "@/src/actions/breadcrumb";
 import { dataPageAction, UPDATE_TITLE } from "@/src/actions/dataPage";
-import { useDispatch } from "react-redux";
+import PostModal from "@/components/Customer/CustomerModal";
+import CustomerDeleteModal from "@/components/Customer/CustomerDeleteModal";
+import CustomerOptionsDrawer from "@/components/Customer/CustomerOptionsDrawer";
+import CustomerEditable from "@/components/Customer/CustomerEditable";
 import {
-  sortByNameDesc,
-  sortByNameAsc,
-  processRequest,
-  processRequestNonReponse,
-  sendMessageAction,
-  getHeaders,
-} from "@/pages/index";
+  EnhancedTableHead,
+  stableSort,
+  getComparator,
+  Order,
+  EnhancedTableToolbar,
+} from "@/components/Customer/CustomerTableUtils";
 
-const styles = {
-  container: {
-    marginTop: "100px",
-    width: "100%",
-    textAlign: "center",
-  },
-  optionsButton: {
-    marginBottom: "50px",
-  },
-  button: {
-    color: "white",
-    fontWeight: "bold",
-    "&:hover": {
-      color: "black",
-      backgroundColor: "white",
-    },
-  },
-};
-
-export default function Customer() {
-  const [objList, setObjList] = useState<TCustomer[]>([]);
+export default function EnhancedTable() {
+  const [rows, setRows] = useState<TCustomer[]>([]);
+  const [rowsBackUp, setRowsBackUp] = useState<TCustomer[]>([]);
+  const [order, setOrder] = React.useState<Order>("asc");
+  const [orderBy, setOrderBy] = React.useState<keyof TCustomer>("name");
+  const [selected, setSelected] = React.useState<readonly number[]>([]);
+  const [page, setPage] = React.useState(0);
+  const [rowsPerPage, setRowsPerPage] = React.useState(10);
   const [loading, setLoading] = useState<boolean>(true);
 
-  const [open, setOpen] = useState(false);
-  const handleOpen = () => setOpen(true);
-  const handleClose = () => setOpen(false);
+  const handleRequestSort = (
+    event: React.MouseEvent<unknown>,
+    property: keyof TCustomer
+  ) => {
+    const isAsc = orderBy === property && order === "asc";
+    setOrder(isAsc ? "desc" : "asc");
+    setOrderBy(property);
+  };
+
+  const handleSelectAllClick = (event: React.ChangeEvent<HTMLInputElement>) => {
+    if (event.target.checked) {
+      const newSelected = rows.map((n) => n.id);
+      setSelected(newSelected);
+      return;
+    }
+    setSelected([]);
+  };
+
+  const router = useRouter();
+  const handleClick = (event: React.MouseEvent<unknown>, obj: TCustomer) => {
+    const selectedIndex = selected.indexOf(obj.id);
+    let newSelected: readonly number[] = [];
+
+    if (selectedIndex === -1) {
+      newSelected = newSelected.concat(selected, obj.id);
+    } else if (selectedIndex === 0) {
+      newSelected = newSelected.concat(selected.slice(1));
+    } else if (selectedIndex === selected.length - 1) {
+      newSelected = newSelected.concat(selected.slice(0, -1));
+    } else if (selectedIndex > 0) {
+      newSelected = newSelected.concat(
+        selected.slice(0, selectedIndex),
+        selected.slice(selectedIndex + 1)
+      );
+    }
+
+    setSelected(newSelected);
+  };
+
+  const handleChangePage = (event: unknown, newPage: number) => {
+    setPage(newPage);
+  };
+
+  const handleChangeRowsPerPage = (
+    event: React.ChangeEvent<HTMLInputElement>
+  ) => {
+    setRowsPerPage(parseInt(event.target.value, 10));
+    setPage(0);
+  };
+
+  const isSelected = (id: number) => selected.indexOf(id) !== -1;
+
+  // Avoid a layout jump when reaching the last page with empty rows.
+  const emptyRows =
+    page > 0 ? Math.max(0, (1 + page) * rowsPerPage - rows.length) : 0;
+
+  const visibleRows = React.useMemo(
+    () =>
+      stableSort(rows, getComparator(order, orderBy)).slice(
+        page * rowsPerPage,
+        page * rowsPerPage + rowsPerPage
+      ),
+    [rows, order, orderBy, page, rowsPerPage]
+  );
 
   const [searchTerm, setSearchTerm] = useState("");
   function handleSearch(
@@ -70,40 +120,45 @@ export default function Customer() {
   ) {
     const term = event.target.value;
     setSearchTerm(term);
-    const filtered = objList.filter((obj) =>
+    const filtered = rowsBackUp.filter((obj) =>
       obj.name.toLowerCase().includes(term.toLowerCase())
     );
-    setSortedList(filtered);
+    setRows(filtered);
   }
 
-  const [sortedList, setSortedList] = useState(objList);
-  const [sortOrder, setSortOrder] = useState("asc");
-  const handleSort = () => {
-    if (sortOrder === "asc") {
-      setSortedList(sortedList.sort(sortByNameDesc));
-      setSortOrder("desc");
-    } else {
-      setSortedList(sortedList.sort(sortByNameAsc));
-      setSortOrder("asc");
-    }
-  };
+  const [open, setOpen] = useState(false);
+  const handleOpen = () => setOpen(true);
+  const handleClose = () => setOpen(false);
 
-  const [checkedList, setCheckedList] = useState<Map<number, boolean>>(
-    new Map()
+  const [isEditable, setIsEditable] = useState(false);
+  const handleIsEditableClose = () => setIsEditable(false);
+
+  const [openToDelete, setOpenToDelete] = useState(false);
+  const handleOpenToDelete = () => setOpenToDelete(true);
+  const handleCloseToDelete = () => setOpenToDelete(false);
+  const [deleteOp, setDeleteOp] = useState<boolean>(
+    selected.length > 0 ? true : false
   );
-  const [deleteOp, setDeleteOp] = useState<boolean>(false);
   function delete_obj() {
     let urls: Array<Promise<Response>> = [];
-    checkedList.forEach(
-      (value: boolean, key: number, map: Map<number, boolean>) => {
-        urls.push(
-          window.fetch(`/api/customer/${key}`, {
-            method: "DELETE",
-            headers: getHeaders(),
-          })
-        );
-      }
-    );
+    if (selected.length == 0) {
+      sendMessageAction(
+        "warning",
+        "No has seleccionado ningún cliente",
+        dispatch
+      );
+      return;
+    }
+
+    selected.forEach((value: number, key: number, map: readonly number[]) => {
+      console.log(`${value} ${key} ${map}`);
+      urls.push(
+        window.fetch(`/api/customer/${value}`, {
+          method: "DELETE",
+          headers: getHeaders(),
+        })
+      );
+    });
     Promise.all(urls).then((responses) => {
       let failed = false;
       for (let response of responses) {
@@ -125,7 +180,7 @@ export default function Customer() {
 
       reload();
       setDeleteOp(false);
-      setCheckedList(new Map());
+      setSelected([]);
     });
   }
 
@@ -145,8 +200,8 @@ export default function Customer() {
       )
       .then((data) => {
         if (data) {
-          setObjList(data);
-          setSortedList(data);
+          setRows(data);
+          setRowsBackUp(data);
         }
         setLoading(false);
       });
@@ -174,92 +229,147 @@ export default function Customer() {
   }, []);
 
   return (
-    <Fragment>
+    <div>
       <PostModal reload={reload} open={open} handleClose={handleClose} />
-      <Grid container spacing={2}>
-        <Grid item>
-          <TextField
-            size="small"
-            label="Buscar por nombre"
-            value={searchTerm}
-            onChange={handleSearch}
-          />
-        </Grid>
-        <Grid item>
-          <Button onClick={handleSort} variant="outlined">
-            Ordenar por nombre (
-            {sortOrder === "asc" ? "ascendente" : "descendente"})
-          </Button>
-        </Grid>
-      </Grid>
-      <Box sx={{ marginY: 2 }} />
-      <Grid container spacing={5} alignItems="flex-end">
-        {sortedList.map((obj) => (
-          <CustomerCard
-            key={obj.id}
-            customer={obj}
-            checkedList={checkedList}
-            setCheckedList={setCheckedList}
-            deleteOp={deleteOp}
-          />
-        ))}
-      </Grid>
-      <Container sx={styles.container}>
-        {loading && (
-          <Box sx={{ width: "100%" }}>
-            <CircularProgress />
-          </Box>
-        )}
-        {!loading && sortedList.length == 0 && (
-          <Typography>
-            {searchTerm.length > 0
-              ? "Ningún cliente coincide con la busqueda"
-              : "Aún no has creado ningún cliente"}
-          </Typography>
-        )}
-      </Container>
 
-      <Box sx={{ position: "fixed", bottom: "32px", right: "32px" }}>
-        <OptionsButton
-          function_1={delete_obj}
-          function_2={() => setDeleteOp(false)}
-          check_or_cancel={deleteOp}
-        >
-          <List>
-            <ListItem key={"new_key"} disablePadding>
-              <ListItemButton onClick={handleOpen}>
-                <ListItemIcon>
-                  <CreateIcon />
-                </ListItemIcon>
-                <ListItemText primary={"Crear nuevo cliente"} />
-              </ListItemButton>
-            </ListItem>
-            <ListItem key={"delete_key"} disablePadding>
-              <ListItemButton onClick={() => setDeleteOp(true)}>
-                <ListItemIcon>
-                  <DeleteIcon />
-                </ListItemIcon>
-                <ListItemText
-                  primary={
-                    "Eliminar cliente (Esto incluye todos las facturas asociadas a este)"
-                  }
-                />
-              </ListItemButton>
-            </ListItem>
-          </List>
-          <Divider />
-          <List>
-            <ListItem key={"exit"} disablePadding>
-              <ListItemButton>
-                <ListItemIcon>
-                  <InboxIcon />
-                </ListItemIcon>
-                <ListItemText primary={"Cerrar"} />
-              </ListItemButton>
-            </ListItem>
-          </List>
-        </OptionsButton>
+      <CustomerDeleteModal
+        open={openToDelete}
+        onClose={handleCloseToDelete}
+        onDelete={delete_obj}
+      />
+
+      <Box sx={{ width: "100%" }}>
+        <Paper sx={{ width: "100%", mb: 2 }}>
+          <EnhancedTableToolbar
+            numSelected={selected.length}
+            handleOpenToDelete={handleOpenToDelete}
+            searchTerm={searchTerm}
+            handleSearch={handleSearch}
+          />
+          <TableContainer>
+            <Table
+              sx={{ minWidth: 750 }}
+              aria-labelledby="tableTitle"
+              size="medium"
+            >
+              <EnhancedTableHead
+                numSelected={selected.length}
+                order={order}
+                orderBy={orderBy}
+                onSelectAllClick={handleSelectAllClick}
+                onRequestSort={handleRequestSort}
+                rowCount={rows.length}
+                deleteOption={deleteOp}
+              />
+              <TableBody>
+                {visibleRows.map((row, index) => {
+                  const isItemSelected = isSelected(row.id);
+                  const labelId = `enhanced-table-checkbox-${index}`;
+
+                  return (
+                    <TableRow
+                      hover
+                      onClick={() =>
+                        !isEditable && router.push(`/customer/${row.id}`)
+                      }
+                      role="checkbox"
+                      aria-checked={isItemSelected}
+                      tabIndex={-1}
+                      key={row.name}
+                      selected={isItemSelected}
+                      sx={{ cursor: "pointer" }}
+                    >
+                      <TableCell padding="checkbox">
+                        {deleteOp && (
+                          <Checkbox
+                            color="primary"
+                            checked={isItemSelected}
+                            onClick={(event) => {
+                              event.stopPropagation();
+                              handleClick(event, row);
+                            }}
+                            inputProps={{
+                              "aria-labelledby": labelId,
+                            }}
+                          />
+                        )}
+                      </TableCell>
+
+                      <TableCell
+                        component="th"
+                        id={labelId}
+                        scope="row"
+                        padding="none"
+                        align="left"
+                      >
+                        <CustomerEditable
+                          customer={row}
+                          isEditable={isEditable}
+                          handleIsEditableClose={handleIsEditableClose}
+                        />
+                      </TableCell>
+                      <TableCell align="center">{row.num_invoices}</TableCell>
+                    </TableRow>
+                  );
+                })}
+                {emptyRows > 0 && (
+                  <TableRow
+                    style={{
+                      height: 53 * emptyRows,
+                    }}
+                  >
+                    <TableCell colSpan={6} />
+                  </TableRow>
+                )}
+
+                {rows.length == 0 && !loading && (
+                  <TableRow
+                    style={{
+                      height: 53,
+                    }}
+                  >
+                    <TableCell colSpan={7} align="center">
+                      Aún no se ha creado ningún cliente
+                    </TableCell>
+                  </TableRow>
+                )}
+                {rows.length == 0 && loading && (
+                  <TableRow>
+                    <TableCell colSpan={7} sx={{ padding: "0 0 15px 0" }}>
+                      <LinearProgress />
+                      <div style={{ textAlign: "center", marginTop: "15px" }}>
+                        Cargando...
+                      </div>
+                    </TableCell>
+                  </TableRow>
+                )}
+              </TableBody>
+            </Table>
+          </TableContainer>
+          <TablePagination
+            rowsPerPageOptions={[5, 10, 25]}
+            component="div"
+            count={rows.length}
+            rowsPerPage={rowsPerPage}
+            page={page}
+            onPageChange={handleChangePage}
+            onRowsPerPageChange={handleChangeRowsPerPage}
+          />
+        </Paper>
       </Box>
-    </Fragment>
+
+      <CustomerOptionsDrawer
+        handleOpenToDelete={handleOpenToDelete}
+        deleteOption={deleteOp}
+        setDeleteOption={setDeleteOp}
+        handleOpen={handleOpen}
+        editable={isEditable}
+        setEditable={setIsEditable}
+        setSelected={setSelected}
+        textCreateOption="Crear nuevo cliente"
+        textUpdateOption="Modificar nombre cliente"
+        textDeleteOption="Eliminar cliente (Esto incluye todos las facturas asociadas a este)"
+      />
+    </div>
   );
 }
